// src/app/admin/models/tvshow.ts
export interface TVShow {
    id: number | undefined;
    name: string | undefined;
    title: string | undefined;
    description: string | undefined;
    imdbrating: string | undefined;
    language: string | undefined;
    genere: string | undefined;
    image: any;
    imageUrl: string | undefined
}